Add files via upload
A data set showing the current employees of a company with resigned or retired employees. This data requires us to find the attrition rate and reasons for employee attrition in the company.
A model is required to be written so as to predict those employees who are likely to leave the company in the near future.